Move Ocr editor to own directory
