+++ frontend/components/Domain/Recipe/RecipeOcrEditorPage/index.ts
@@ -1,0 +1,3 @@
+import RecipeOcrEditorPage from "./RecipeOcrEditorPage.vue";
+
+export default RecipeOcrEditorPage;
